requirements.txt
# PyTorch 2.6.0 with CUDA 11.8
torch
torchaudio
torchvision
# HuggingFace ecosystem
transformers
accelerate
diffusers
# Qt & GUI
PySide6
python-dotenv